Move RequestItem into request_items module
